+++ app/(footer-pages)/terms/page.tsx
@@ -1,28 +1,36 @@
-import React from 'react';
+import React from "react";
 
 export default function TermsPage() {
-    return (
-      <div className="max-w-3xl mx-auto p-4">
-        <h1 className="text-4xl font-bold mb-4">Terms of Service</h1>
-        <p className="mb-2">
-          Welcome to Elevra. These Terms of Service (&quot;Terms&quot;) govern your access to and use of our website and services.
-        </p>
-        <h2 className="text-2xl font-semibold mt-6 mb-2">Acceptance of Terms</h2>
-        <p className="mb-2">
-          By accessing or using Elevra, you agree to be bound by these Terms. If you do not agree, please do not use our services.
-        </p>
-        <h2 className="text-2xl font-semibold mt-6 mb-2">Use of Service</h2>
-        <p className="mb-2">
-          Elevra is provided for informational purposes only. We strive to ensure accuracy and reliability, but cannot guarantee that all information is complete or error-free.
-        </p>
-        <h2 className="text-2xl font-semibold mt-6 mb-2">Changes to Terms</h2>
-        <p className="mb-2">
-          We reserve the right to update or modify these Terms at any time. Your continued use of the service constitutes acceptance of the new Terms.
-        </p>
-        <h2 className="text-2xl font-semibold mt-6 mb-2">Contact Us</h2>
-        <p>
-          If you have any questions about these Terms, please contact us at <a href="mailto:elevracommunity@gmail.com" className="text-blue-500">elevracommunity@gmail.com</a>.
-        </p>
-      </div>
-    );
-  }
+  return (
+    <div className="max-w-3xl mx-auto p-4">
+      <h1 className="text-4xl font-bold mb-4">Terms of Service</h1>
+      <p className="mb-2">
+        Welcome to Elevra. These Terms of Service (&quot;Terms&quot;) govern
+        your access to and use of our website and services.
+      </p>
+      <h2 className="text-2xl font-semibold mt-6 mb-2">Acceptance of Terms</h2>
+      <p className="mb-2">
+        By accessing or using Elevra, you agree to be bound by these Terms. If
+        you do not agree, please do not use our services.
+      </p>
+      <h2 className="text-2xl font-semibold mt-6 mb-2">Use of Service</h2>
+      <p className="mb-2">
+        We strive to ensure accuracy and reliability, but cannot guarantee that
+        all information is complete or error-free.
+      </p>
+      <h2 className="text-2xl font-semibold mt-6 mb-2">Changes to Terms</h2>
+      <p className="mb-2">
+        We reserve the right to update or modify these Terms at any time. Your
+        continued use of the service constitutes acceptance of the new Terms.
+      </p>
+      <h2 className="text-2xl font-semibold mt-6 mb-2">Contact Us</h2>
+      <p>
+        If you have any questions about these Terms, please contact us at{" "}
+        <a href="mailto:elevracommunity@gmail.com" className="text-blue-500">
+          elevracommunity@gmail.com
+        </a>
+        .
+      </p>
+    </div>
+  );
+}
