+++ tsconfig.json
@@ -22,6 +22,6 @@
       "@/*": ["./*"], 
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/(layout)/LayoutClient.tsx"],
   "exclude": ["node_modules"]
 }
